Refactor import statements and improve balance calculations in Tiles NFT tests
- Cleaned up import statements across multiple test files for better organization and readability.
- Updated balance calculations in the `mint.rs` test to remove unnecessary type casting, enhancing clarity.
- Improved assertions in the `setup.rs` test to use more descriptive checks for contract address validity.

These changes enhance the maintainability and clarity of the testing framework for the Tiles NFT contract, facilitating future development.

+++ tests/common/users/roles.rs
@@ -36,14 +36,14 @@ impl UserConfig {
     pub fn buyer() -> Self {
         Self {
             role: UserRole::Buyer,
-            initial_balance: MINT_PRICE as u128 * 10,
+            initial_balance: MINT_PRICE * 10,
         }
     }
 
     pub fn whale() -> Self {
         Self {
             role: UserRole::Whale,
-            initial_balance: MINT_PRICE as u128 * 100,
+            initial_balance: MINT_PRICE * 100,
         }
     }
